requirements.txt
discord.py
requests
youtube_dl